Add new application context, utility classes, dispatcher servlet, and query builder; refactor ConnectionPool to use AnnotationConfigApplicationContext
This commit implements a new context that can bootstrap the application based on annotation metadata. It also includes utility classes such as ConcurrentWeakHashMap that can be used to cache queries, as well as a DispatcherServlet and ControllerFactory. Additionally, the ConnectionPool singleton implementation has been refactored to become a singleton in the AnnotationConfigApplicationContext container.

Finally, QueryBuilder and related classes have been added to create and cache queries. These changes should improve performance and maintainability